Revert to python:3.10-slim, tidy repo
Pillow and regex have wheel packages available for CPython 3.10, which
means that compiling these packages from source is no longer required.
This not only reduces the time to build the image, but also means that
all of the build-time dependencies could be removed, and the image is
much smaller than the alpine-based image.

Repo tidyup also performed, broad changes:
- Move Python files into `src` subdirectory
- Finish README
- Move images into `images` subdirectory
- Allow specifying hostname via `LEDSERVER_HOST` environment variable
- Small error detection change to check bytes were sent correctly